% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hybrid_recursive_mc.R
\name{hybrid_recursive_mc}
\alias{hybrid_recursive_mc}
\title{Hybrid-recursive Outlier Removal Procedure with Moving Criterion}
\usage{
hybrid_recursive_mc(exp_cell)
}
\arguments{
\item{exp_cell}{Numeric vector on which the outlier removal method takes place.}
}
\value{
Mean of \code{exp_cell} after removing outliers.
}
\description{
Hybrid-recursive outlier removal procedure with moving criterion according to Van Selst & Jolicoeur (1994).
}
\references{
Grange, J.A. (2015). trimr: An implementation of common response time trimming methods. R Package Version 1.0.0. \url{http://cran.r-project.org/web/packages/trimr/}

Selst, M. V., & Jolicoeur, P. (1994). A solution to the effect of sample size on outlier elimination. \emph{The quarterly journal of experimental psychology, 47}(3), 631-650.
}

