% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linear_interpolation.R
\docType{data}
\name{linear_interpolation}
\alias{linear_interpolation}
\title{Criterion cutoffs for the non-recursive and modified recursive outlier removal procedures with moving criterion
according to procedures described by Van Selst & Jolicoeur (1994), Table 4, page 642.}
\format{A data frame with 101 rows and 3 columns:
\describe{
  \item{\code{sample_size}}{sample size, in numerals}
  \item{\code{non_recursive}}{criterion cutoffs for non-recursive with moving criterion procedure}
  \item{\code{modified_recursive}}{criterion cutoffs for modified recursive with moving criterion procedure}
}}
\usage{
\code{data(linear_interpolation)}
}
\description{
A table containing criterion cutoffs for non-recursive and modified recursive outlier removal procedures with
moving criterion according to sample size.
}
\references{
Grange, J.A. (2015). trimr: An implementation of common response time trimming methods. R Package Version 1.0.0. \url{http://cran.r-project.org/web/packages/trimr/}

Selst, M. V., & Jolicoeur, P. (1994). A solution to the effect of sample size on outlier eliminatio
\emph{The quarterly journal of experimental psychology, 47}(3), 631-650.
}
\keyword{datasets}

