% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stroopdata.R
\docType{data}
\name{stroopdata}
\alias{stroopdata}
\title{Reaction-times and accuracy for color naming in a Stroop task (e.g., Stroop, 1935).}
\format{A data frame with 5401 rows and 10 columns:
\describe{
  \item{\code{subject}}{case identifier, in numerals}
  \item{\code{block}}{percent of congruent target_type trials in a block.
   1 means 80 percent congruent, 2 means 20 percent congruent}
  \item{\code{age}}{age of subject, in integers}
  \item{\code{gender}}{gender of subject, in integers. 1 means male, 2 means female}
  \item{\code{order}}{order of blocks, in integers.
   1 means subject did 80 percent congruent block first and 20 percent congruent block second.
   2 means subject did 20 percent congruent block first and 80 percent congruent block second.}
  \item{\code{font_size}}{font size of the stimulus, in integers}
  \item{\code{trial_num}}{trial number, in integers}
  \item{\code{target_type}}{type of stimulus for a given trial. 1 means congruent stimulus,
  2 means incongruent stimulus}
  \item{\code{rt}}{reaction time, in milliseconds}
  \item{\code{ac}}{accuracy, 1 means correct, 0 means incorrect}
}}
\usage{
\code{data(stroopdata)}
}
\description{
A dataset containing reaction-times, accuracy, and other attributes of 5400
experimental trials.
}
\examples{
data(stroopdata)
head(stroopdata)
}
\references{
Stroop, J. R. (1935). Studies of interference in serial verbal reactions. \emph{Journal of experimental psychology, 18}(6), 643.
}
\keyword{datasets}

